Added API for doctor registration

See merge request nmr/abdm-nmr!30

+++ src/main/java/in/gov/abdm/nmr/db/sql/domain/state_medical_council/IStateMedicalCouncilRepository.java
@@ -1,7 +1,9 @@
 package in.gov.abdm.nmr.db.sql.domain.state_medical_council;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IStateMedicalCouncilRepository extends JpaRepository<StateMedicalCouncil, Long> {
+public interface IStateMedicalCouncilRepository extends JpaRepository<StateMedicalCouncil, BigInteger> {
 
 }
